refactor: centralize copy functionality and enhance CodeBlock component
- Introduced a new utility function for copying text to the clipboard, improving code reusability across components.
- Updated CodeBlock and ToolCall components to utilize the new copy function, streamlining the copy process and enhancing error handling.
- Added a timer to manage the copied state in CodeHeader for better user feedback.
