+++ packages/spark-chat/components/AgentScopeRuntimeWebUI/core/AgentScopeRuntime/Response/Actions.tsx
@@ -2,7 +2,8 @@ import { SparkCopyLine, SparkReplaceLine } from "@agentscope-ai/icons";
 import { IAgentScopeRuntimeResponse } from "../types";
 import AgentScopeRuntimeResponseBuilder from "./Builder";
 import { Bubble } from "@agentscope-ai/chat";
-import { copy, Tooltip } from "@agentscope-ai/design";
+import { Tooltip } from "@agentscope-ai/design";
+import { copy } from "../../../../Util/copy";
 import compact from 'lodash/compact';
 import { emit } from "../../Context/useChatAnywhereEventEmitter";
 import { useChatAnywhereOptions } from "../../Context/ChatAnywhereOptionsContext";
